chore: refactor some dangerous imports
