+++ src/app/components/stream-controller/stream-controller.component.ts
@@ -3,8 +3,9 @@ import { Clipboard } from '@angular/cdk/clipboard';
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { Stream } from '../../core';
-import { InsightsService, LoggerService, RuntimeService, StreamBuilderService } from '../../services';
+import { StreamLike } from '../../core';
+import { LoggerService } from '../../logger.service';
+import { InsightsService, RuntimeService, StreamBuilderService } from '../../services';
 import { StreamOptionsComponent } from '../stream-options/stream-options.component';
 
 @Component({
@@ -13,7 +14,7 @@ import { StreamOptionsComponent } from '../stream-options/stream-options.compone
   styleUrls: ['./stream-controller.component.scss']
 })
 export class StreamControllerComponent implements OnInit {
-  @Input() stream: Stream | undefined;
+  @Input() stream: StreamLike | undefined;
   @Input() canRemoveSource$ = of(false);
   @Input() hasOptions$ = of(false);
