+++ src/app/core/types/stream.ts
@@ -1,3 +1,4 @@
+import { Observable } from 'rxjs';
 import { Notification } from './notification';
 
 export interface StreamMarbles {
@@ -19,4 +20,22 @@ export interface StreamConfig {
   dy: number;
   offset: number;
   frames: { small: number; large: number };
+}
+
+export interface StreamLike {
+  dx: number;
+  dy: number;
+  offset: number;
+  frames: number;
+
+  entities$: Observable<{ [id: string]: StreamNode }>;
+  nodes$: Observable<StreamNode[]>;
+  next$: Observable<StreamNode[]>;
+  terminate$: Observable<StreamNode | null>;
+  nodesToRender$: Observable<StreamNode[]>;
+  marbles$: Observable<StreamMarbles>;
+}
+
+export interface InputStreamLike extends StreamLike {
+  updateNode(update: Partial<StreamNode> & { id: string }): void;
 }
