+++ src/app/examples/catch-error.spec.ts
@@ -3,7 +3,7 @@ import { MockService } from 'ng-mocks';
 import { first, map, of, tap } from 'rxjs';
 import { TestBed } from '@angular/core/testing';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { InputStream } from '../core';
+import { InputStreamLike } from '../core';
 import { ExecutorService, RuntimeService, StreamBuilderService } from '../services';
 import { CatchErrorExample } from './catch-error';
 
@@ -39,7 +39,7 @@ describe('CatchErrorExample', () => {
     const outputValues = { A: '1', B: '2', C: 'a' };
 
     let code: string;
-    let inputStreams: InputStream[];
+    let inputStreams: InputStreamLike[];
 
     beforeEach(() => {
       streamBuilderSvc.setFrames('large');
@@ -91,7 +91,7 @@ describe('CatchErrorExample', () => {
     const outputValues = { A: '1', B: '2', C: 'a' };
 
     let code: string;
-    let inputStreams: InputStream[];
+    let inputStreams: InputStreamLike[];
 
     beforeEach(() => {
       streamBuilderSvc.setFrames('small');
