+++ src/app/examples/concat-map.spec.ts
@@ -3,7 +3,7 @@ import { MockService } from 'ng-mocks';
 import { first, map, of, tap } from 'rxjs';
 import { TestBed } from '@angular/core/testing';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { InputStream } from '../core';
+import { InputStreamLike } from '../core';
 import { ExecutorService, RuntimeService, StreamBuilderService } from '../services';
 import { ConcatMapExample } from './concat-map';
 
@@ -40,7 +40,7 @@ describe('ConcatMapExample', () => {
     const outputValues = { A: '1a', B: '1b', C: '2a', D: '2b', E: '3a', F: '3b', G: '4a', H: '4b' };
 
     let code: string;
-    let inputStreams: InputStream[];
+    let inputStreams: InputStreamLike[];
 
     beforeEach(() => {
       streamBuilderSvc.setFrames('large');
@@ -98,7 +98,7 @@ describe('ConcatMapExample', () => {
     const outputValues = { A: '1a', B: '1b', C: '2a', D: '2b' };
 
     let code: string;
-    let inputStreams: InputStream[];
+    let inputStreams: InputStreamLike[];
 
     beforeEach(() => {
       streamBuilderSvc.setFrames('small');
