+++ src/app/examples/merge.spec.ts
@@ -3,7 +3,7 @@ import { MockService } from 'ng-mocks';
 import { first, map, of, tap } from 'rxjs';
 import { TestBed } from '@angular/core/testing';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { InputStream } from '../core';
+import { InputStreamLike } from '../core';
 import { ExecutorService, RuntimeService, StreamBuilderService } from '../services';
 import { MergeExample } from './merge';
 
@@ -40,7 +40,7 @@ describe('MergeExample', () => {
     const outputValues = { A: 'a', B: '1', C: '2', D: 'b', E: '3' };
 
     let code: string;
-    let inputStreams: InputStream[];
+    let inputStreams: InputStreamLike[];
 
     beforeEach(() => {
       streamBuilderSvc.setFrames('large');
@@ -98,7 +98,7 @@ describe('MergeExample', () => {
     const outputValues = { A: 'a', B: '1', C: '2', D: 'b', E: '3' };
 
     let code: string;
-    let inputStreams: InputStream[];
+    let inputStreams: InputStreamLike[];
 
     beforeEach(() => {
       streamBuilderSvc.setFrames('small');
