+++ src/app/examples/switch-map.spec.ts
@@ -3,7 +3,7 @@ import { MockService } from 'ng-mocks';
 import { first, map, of, tap } from 'rxjs';
 import { TestBed } from '@angular/core/testing';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { InputStream } from '../core';
+import { InputStreamLike } from '../core';
 import { ExecutorService, RuntimeService, StreamBuilderService } from '../services';
 import { SwitchMapExample } from './switch-map';
 
@@ -40,7 +40,7 @@ describe('SwitchMapExample', () => {
     const outputValues = { A: '1a', B: '2a', C: '2b', D: '3a', E: '4a', F: '4b' };
 
     let code: string;
-    let inputStreams: InputStream[];
+    let inputStreams: InputStreamLike[];
 
     beforeEach(() => {
       streamBuilderSvc.setFrames('large');
@@ -98,7 +98,7 @@ describe('SwitchMapExample', () => {
     const outputValues = { A: '1a', B: '2a', C: '3a', D: '3b' };
 
     let code: string;
-    let inputStreams: InputStream[];
+    let inputStreams: InputStreamLike[];
 
     beforeEach(() => {
       streamBuilderSvc.setFrames('small');
