+++ src/app/services.ts
@@ -1,6 +1,5 @@
 export * from './services/executor.service';
 export * from './services/insights.service';
 export * from './services/local-storage.service';
-export * from './services/logger.service';
 export * from './services/runtime.service';
 export * from './services/stream.builder';
